refactor: break out hono rpcs into separate handler files
- Create handlers/ directory with factory pattern approach
- Separate get_game_state and ship_fly_to into individual files
- Maintain type safety with explicit dependency injection
- Clean handler composition in main index.ts
- Each handler exports factory function accepting state dependency
