+++ apps/server/src/handlers/get-game-state.ts
@@ -1,0 +1,11 @@
+import type { ApiResponse } from '@space/api'
+import type { WorldState } from '../game/world-state.js'
+
+export const createGetGameStateHandler = (state: WorldState) => {
+  return async (): Promise<ApiResponse> => {
+    return {
+      success: true,
+      data: state,
+    }
+  }
+}
