+++ apps/server/src/handlers/index.ts
@@ -1,0 +1,2 @@
+export { createGetGameStateHandler } from './get-game-state.js'
+export { createShipFlyToHandler } from './ship-fly-to.js'
