+++ packages/game/src/world-state.ts
@@ -3,11 +3,7 @@ import { PlanetSchema } from './game/planet.js'
 import { ShipSchema } from './game/ship.js'
 import { StarSchema } from './game/star.js'
 
-export const WorldEntitySchema = z.union([
-  StarSchema,
-  PlanetSchema,
-  ShipSchema,
-])
+export const WorldEntitySchema = z.union([StarSchema, PlanetSchema, ShipSchema])
 
 export const WorldStateSchema = z.object({
   entities: z.array(WorldEntitySchema),
